* Updated entities to replace decimal properties to double properties ** Removed decimal casts where needed. ** Removed DecimalConversionExtensions as they are now obsolete. Related to Issue [WTI-691]
git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@4340 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: db194dd8c6e2e2f37c9fd380254059fc1d5d37ae
Former-commit-id: 854c8aa94eb580c8a369f297e7a1ae8d91b5b290

+++ Application/Ringtoets/src/Application.Ringtoets.Storage/DbContext/ClosingStructureSectionResultEntity.cs
@@ -38,8 +38,8 @@ namespace Application.Ringtoets.Storage.DbContext
         public long ClosingStructureSectionResultEntityId { get; set; }
         public long FailureMechanismSectionEntityId { get; set; }
         public byte LayerOne { get; set; }
-        public Nullable<decimal> LayerTwoA { get; set; }
-        public Nullable<decimal> LayerThree { get; set; }
+        public Nullable<double> LayerTwoA { get; set; }
+        public Nullable<double> LayerThree { get; set; }
     
         public virtual FailureMechanismSectionEntity FailureMechanismSectionEntity { get; set; }
     }
